テーブル定義書の追加、DDLとEntityの微修正

+++ src/SkillManager/src/main/java/com/umayoryo/skill/manager/service/engineer/EngineerBean.java
@@ -7,13 +7,13 @@ import lombok.Data;
  */
 @Data
 public class EngineerBean {
-    
+
     /** エンジニアID. */
-    private int engineerId;
-    
+    private long engineerId;
+
     /**エンジニア名. */
     private String engineerName;
-    
+
     /** エンジニアスキル名. */
     private String skillName;
 }
